Cleaned balances.py of static methods and moved it to types package, adjusted balances_gateway imports accordingly and added more inline comments.
