+++ trolley/__init__.py
@@ -8,7 +8,7 @@ from trolley.types.batch import Batch
 from trolley.types.batch_summary import BatchSummary
 from trolley.types.payment import Payment
 from trolley.types.offline_payment import OfflinePayment
-from trolley.balances import Balances
+from trolley.types.balances import Balances
 from trolley.types.invoice import Invoice
 from trolley.types.invoice_line import InvoiceLine
 from trolley.types.invoice_payment import InvoicePayment
