+++ trolley/types/balances.py
@@ -1,6 +1,3 @@
-from trolley.configuration import Configuration
-from trolley.gateway import Gateway
-
 class Balances:
     """
     A class that facilitates Client requests to
